check if file exists otherwise ignore

+++ labs/drill/api-primitives/teardown.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 kubectl delete --ignore-not-found pods apricot -n pineapple
-kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-api-primitives.yaml
-echo "Teardown complete"
+if [ -e ~/mycode/yaml/ctce-drill-api-primitives.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-api-primitives.yaml
+else
+  echo "file does not exist"
+fi
