+++ labs/drill/probes/teardown.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-probes.yaml
-kubectl delete --ignore-not-found -f ~/plumpod.yaml
+if [ -e ~/plumpod.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/plumpod.yaml
+else
+  echo "file does not exist"
+fi
 echo "Teardown complete"
