+++ labs/drill/secrets/teardown.sh
@@ -1,5 +1,17 @@
 #!/bin/bash
 kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-secrets.yaml
-kubectl delete --ignore-not-found -f ~/juicysecret.yaml
-kubectl delete --ignore-not-found -f ~/kiwi-secret-pod.yaml
+if [ -e ~/juicysecret.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/juicysecret.yaml
+else
+  echo "file does not exist"
+fi
+if [ -e ~/kiwi-secret-pod.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/kiwi-secret-pod.yaml
+else
+  echo "file does not exist"
+fi
 echo "Teardown complete"
