+++ labs/drill/securitycontexts/teardown.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-security-contexts.yaml
-kubectl delete --ignore-not-found -f ~/gold-bar.yaml
+if [ -e ~/gold-bar.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/gold-bar.yaml
+else
+  echo "file does not exist"
+fi
 echo "Teardown complete"
