+++ labs/drill/serviceaccounts/teardown.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
 kubectl delete --ignore-not-found -f ~/mycode/yaml/ctce-drill-service-accounts.yaml
-kubectl delete --ignore-not-found -f ~/banana-deployment.yaml
+if [ -e ~/banana-deployment.yaml ] 
+then
+  echo "deleting"
+  kubectl delete --ignore-not-found -f ~/banana-deployment.yaml
+else
+  echo "file does not exist"
+fi
 echo "Teardown complete"
